PR comments applied
